+++ src/context.rs
@@ -8,7 +8,7 @@ use futures::{
 };
 use pollable_map::optional::Optional;
 
-use crate::{Channel, p2p, p2p::MultiaddrExt};
+use crate::{p2p, p2p::MultiaddrExt};
 
 use crate::repo::{Repo, RepoEvent};
 use crate::{IpfsEvent, config::BOOTSTRAP_NODES};
@@ -32,8 +32,9 @@ pub struct IpfsContext {
     pub repo: Repo<DefaultStorage>,
     pub bootstraps: HashSet<Multiaddr>,
     pub find_peer_identify: HashMap<PeerId, Vec<oneshot::Sender<anyhow::Result<Info>>>>,
-    pub relay_listener: HashMap<PeerId, Vec<Channel<()>>>,
     pub discovery_tx: Option<Sender<Cid>>,
+    pub gateway_tx: Option<Sender<Cid>>,
+    pub router_tx: Option<Sender<Cid>>,
 }
 
 impl Default for IpfsContext {
@@ -43,8 +44,9 @@ impl Default for IpfsContext {
             repo: Repo::new_memory(),
             bootstraps: Default::default(),
             find_peer_identify: Default::default(),
-            relay_listener: Default::default(),
             discovery_tx: None,
+            gateway_tx: None,
+            router_tx: None,
         }
     }
 }
@@ -56,8 +58,9 @@ impl IpfsContext {
             repo: repo.clone(),
             bootstraps: Default::default(),
             find_peer_identify: Default::default(),
-            relay_listener: Default::default(),
             discovery_tx: None,
+            gateway_tx: None,
+            router_tx: None,
         }
     }
 }
@@ -298,83 +301,6 @@ impl IpfsContext {
 
                 let _ = ret.send(Ok(rets));
             }
-            IpfsEvent::AddRelay(peer_id, addr, tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_mut() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                relay.add_address(peer_id, addr);
-
-                let _ = tx.send(Ok(()));
-            }
-            IpfsEvent::RemoveRelay(peer_id, addr, tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_mut() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                relay.remove_address(peer_id, addr);
-
-                let _ = tx.send(Ok(()));
-            }
-            IpfsEvent::EnableRelay(Some(peer_id), tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_mut() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                relay.select(peer_id);
-
-                self.relay_listener.entry(peer_id).or_default().push(tx);
-            }
-            IpfsEvent::EnableRelay(None, tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_mut() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                let Some(peer_id) = relay.random_select() else {
-                    let _ = tx.send(Err(anyhow::anyhow!(
-                        "No relay was selected or was unavailable"
-                    )));
-                    return;
-                };
-
-                self.relay_listener.entry(peer_id).or_default().push(tx);
-            }
-            IpfsEvent::DisableRelay(peer_id, tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_mut() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-                relay.disable_relay(peer_id);
-
-                let _ = tx.send(Ok(()));
-            }
-            IpfsEvent::ListRelays(tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_ref() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                let list = relay
-                    .list_relays()
-                    .map(|(peer_id, addrs)| (*peer_id, addrs.clone()))
-                    .collect();
-
-                let _ = tx.send(Ok(list));
-            }
-            IpfsEvent::ListActiveRelays(tx) => {
-                let Some(relay) = self.custom_behaviour(swarm).relay_manager.as_ref() else {
-                    let _ = tx.send(Err(anyhow::anyhow!("Relay is not enabled")));
-                    return;
-                };
-
-                let list = relay.list_active_relays();
-
-                let _ = tx.send(Ok(list));
-            }
         }
     }
 
@@ -388,6 +314,11 @@ impl IpfsContext {
         match event {
             RepoEvent::WantBlock(cids, peers, timeout) => {
                 let Some(bs) = custom.bitswap.as_mut() else {
+                    if let Some(tx) = self.gateway_tx.as_mut() {
+                        for cid in cids {
+                            let _ = tx.try_send(cid);
+                        }
+                    }
                     return;
                 };
                 bs.gets(cids, &peers, timeout);